+++ src/tournament.c
@@ -8,14 +8,14 @@ TournametData DeepCopyTournametData(TournametData tournament) {
       .count = tournament.count,
       .items = malloc(sizeof *tournament.items * tournament.count),
   };
-  for (size_t i = 0; i < tournament.count; i++) {
+  for (unsigned i = 0; i < tournament.count; i++) {
     new_tournament.items[i] = DeepCopyGameLog(tournament.items[i]);
   }
   return tournament;
 }
 
 void FreeInnerTournametData(TournametData tournament) {
-  for (size_t i = 0; i < tournament.count; i++) {
+  for (unsigned i = 0; i < tournament.count; i++) {
     FreeInnerGameLog(tournament.items[i]);
   }
   nob_da_free(tournament);
@@ -23,7 +23,7 @@ void FreeInnerTournametData(TournametData tournament) {
 
 #define GetBotName(idx)                                                        \
   (bots.items[idx].name ? bots.items[idx].name                                 \
-                        : nob_temp_sprintf("Player %zu", idx + 1))
+                        : nob_temp_sprintf("Player %u", idx + 1))
 
 TournametData RunTournament(const char *map_file_path, BotsDA bots) {
   if (bots.count < 3) {
@@ -36,8 +36,8 @@ TournametData RunTournament(const char *map_file_path, BotsDA bots) {
   playing_bots.capacity = playing_bots.count;
   playing_bots.items = malloc(sizeof *playing_bots.items * playing_bots.count);
 
-  for (size_t p1_idx = 0; p1_idx < bots.count - 1; p1_idx++) {
-    for (size_t p2_idx = p1_idx + 1; p2_idx < bots.count; p2_idx++) {
+  for (unsigned p1_idx = 0; p1_idx < bots.count - 1; p1_idx++) {
+    for (unsigned p2_idx = p1_idx + 1; p2_idx < bots.count; p2_idx++) {
       playing_bots.items[0] = bots.items[p1_idx];
       playing_bots.items[1] = bots.items[p2_idx];
       nob_log(NOB_INFO, "Running match between %s and %s", GetBotName(p1_idx),
@@ -52,7 +52,7 @@ TournametData RunTournament(const char *map_file_path, BotsDA bots) {
       nob_minimal_log_level = NOB_INFO;
 
       GameLog game_log_copy = DeepCopyGameLog(state.game_log);
-      size_t winner_idx = game_log_copy.winning_bot == 0 ? p1_idx : p2_idx;
+      unsigned winner_idx = game_log_copy.winning_bot == 0 ? p1_idx : p2_idx;
       nob_log(NOB_INFO, "%s won.", GetBotName(winner_idx));
 
       nob_da_append(&tournament, game_log_copy);
